Chrome extension! (#15)
* Step 1

This commit added a new webpack config which will generate bundled code for the extension. This code will use _ the method of hacking for an extension. In step 2 we should only use the src/hack.tsx file if process.env.EXTENSION is false.

* Step 2

* Step 3

I only added Conjure Cubes as a extension only hack.

* Update extension-bundle.js